// src/app/apps/bizzcalc/models/business-details.interface.ts
export interface IBusinessDetails {
  ID: number;
  CompanyName: string;
  ProjectMonth: string;
  ProjectYear: string;
  ProjectStatus: string;
  StartMonth: string;
  Day1: number;
  Day2: number;
  Day3: number;
  Day4: number;
  Day5: number;
  Day6: number;
  Day7: number;
  WorkDaysForecast: number;
  InventoryDays: string;
  PayablesDays: string;
  ReceivablesDays: string;
}